Struct graphql_client_codegen::GraphQLClientCodegenOptions

pub struct GraphQLClientCodegenOptions {
    pub mode: CodegenMode,
    pub operation_name: Option<String>,
    pub struct_name: Option<String>,
    /* private fields */
}

Used to configure code generation.

Fields
mode: CodegenMode

Which context is this code generation effort taking place.

operation_name: Option<String>

Name of the operation we want to generate code for. If it does not match, we use all queries.

struct_name: Option<String>

The name of implementation target struct.

Implementations

impl GraphQLClientCodegenOptions

pub fn new(mode: CodegenMode) -> GraphQLClientCodegenOptions

Creates an empty options object with default params. It probably wants to be configured.

pub fn set_query_file(&mut self, path: PathBuf)

A path to a file to include in the module to force Cargo to take into account changes in the query files when recompiling.

pub fn variables_derives(&self) -> Option<&str>

Comma-separated list of additional traits we want to derive for variables.

pub fn set_variables_derives(&mut self, variables_derives: String)

Comma-separated list of additional traits we want to derive for variables.

pub fn all_variable_derives(&self) -> impl Iterator<Item = &str>

All the variable derives to be rendered.

pub fn all_response_derives(&self) -> impl Iterator<Item = &str>

Traits we want to derive for responses.

pub fn additional_response_derives(&self) -> impl Iterator<Item = &str>

Additional traits we want to derive for responses.

pub fn set_response_derives(&mut self, response_derives: String)

Comma-separated list of additional traits we want to derive for responses.

pub fn set_deprecation_strategy(
&mut self,
deprecation_strategy: DeprecationStrategy
)

The deprecation strategy to adopt.

pub fn set_module_visibility(&mut self, visibility: Visibility)

Target module visibility.

pub fn set_struct_name(&mut self, struct_name: String)

The name of implementation target struct.

pub fn set_operation_name(&mut self, operation_name: String)

Name of the operation we want to generate code for. If none is selected, it means all operations.

pub fn schema_file(&self) -> Option<&Path>

A path to a file to include in the module to force Cargo to take into account changes in the schema files when recompiling.

pub fn query_file(&self) -> Option<&Path>

A path to a file to include in the module to force Cargo to take into account changes in the query files when recompiling.

pub fn set_struct_ident(&mut self, ident: Ident)

The identifier to use when referring to the struct implementing GraphQLQuery, if any.

pub fn struct_ident(&self) -> Option<&Ident>

The identifier to use when referring to the struct implementing GraphQLQuery, if any.

pub fn set_normalization(&mut self, norm: Normalization)

Set the normalization mode for the generated code.

pub fn normalization(&self) -> &Normalization

The normalization mode for the generated code.

pub fn custom_scalars_module(&self) -> Option<&Path>

Get the custom scalar definitions module

pub fn set_custom_scalars_module(&mut self, module: Path)

Set the custom scalar definitions module

pub fn extern_enums(&self) -> &[String]

Get the externally defined enums type names

pub fn set_extern_enums(&mut self, enums: Vec<String>)

Set the externally defined enums type names

pub fn set_fragments_other_variant(&mut self, fragments_other_variant: bool)

Set the graphql client codegen options’s fragments other variant.

pub fn fragments_other_variant(&self) -> &bool

Get a reference to the graphql client codegen options’s fragments other variant.